+++ packages/ts-components/src/components/newsletter-puff/newsletter-puff-link/NewsletterPuffLink.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import Link from '@times-components/link';
 
-import { TrackingContextProvider } from '../../helpers/tracking/TrackingContextProvider';
+import { TrackingContextProvider } from '../../../helpers/tracking/TrackingContextProvider';
 
-import {  InpPreferencesText } from './styles';
+import { InpPreferencesText } from '../styles';
 
 export const NewsletterPuffLink = ({ onPress }: any) => {
   return (
@@ -32,17 +32,14 @@ export const NewsletterPuffLink = ({ onPress }: any) => {
             });
         };
         return (
-
           <Link
             url="https://home.thetimes.co.uk/myNews"
             onPress={() => handlePress()}
-         
           >
-           
-            <InpPreferencesText ref={intersectObserverRef} >Manage preferences here</InpPreferencesText>
-     
+            <InpPreferencesText ref={intersectObserverRef}>
+              Manage preferences here
+            </InpPreferencesText>
           </Link>
-
         );
       }}
     </TrackingContextProvider>
